backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
